Fix COMMAND_LIST_DRAW_VERTEX_BUFFER_TOO_SMALL in ChunksScene
This fixes the following error:

D3D12 WARNING: ID3D12CommandList::DrawInstanced: Vertex Buffer at the input vertex slot 0 is not big enough for what the Draw*() call expects to traverse. This is OK, as reading off the end of the Buffer is defined to return 0. However the developer probably did not intend to make use of this behavior.  [ EXECUTION WARNING #210: COMMAND_LIST_DRAW_VERTEX_BUFFER_TOO_SMALL]

原先无论画人物还是球，都会发送DrawIndexed(11988, 1)，但是人物实际只有3498个顶点。
这是因为DX12ChunksScene::chunk是球。
把chunk改为c，该错误消失。